requirements.txt
asgiref==3.6.0
Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
